Move installations actions to the reducer
