+++ tests/unit/amo/components/TestInstallButtonWrapper.js
@@ -6,7 +6,7 @@ import GetFirefoxButton, {
 import InstallButtonWrapper, {
   InstallButtonWrapperBase,
 } from 'amo/components/InstallButtonWrapper';
-import { setInstallState } from 'core/actions/installations';
+import { setInstallState } from 'core/reducers/installations';
 import AMInstallButton from 'core/components/AMInstallButton';
 import { CLIENT_APP_FIREFOX, INSTALLED, UNKNOWN } from 'core/constants';
 import { createInternalAddon } from 'core/reducers/addons';
